littlelemon/requirements.txt
-i https://pypi.org/simple
asgiref==3.7.2 ; python_version >= '3.7'
django==4.2.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
pyjwt==2.8.0 ; python_version >= '3.7'
pytz==2023.3
sqlparse==0.4.4 ; python_version >= '3.5'
tzdata==2023.3 ; sys_platform == 'win32'
